Fix #22691: Add PMTiles plugin
This should be fairly easy to port to JOSM core, if there is demand from tile
providers to use PMTiles.

Caveats: There is a few classes copied from JOSM core, since I needed some things
to be visible that were not, and in some instances, the AbstractClass had some
methods that should be hoisted to the interface -- the AbstractClass in question
was written prior to our move to Java 8.